refactor: address refactor related feedback, inc offline
